Revert "fix tests"
This reverts commit ecc7d1de14edb655278d974d9c10accf3699942f.
