+++ src/Routers/FunctionsRouter.js
@@ -2,8 +2,7 @@
 
 const { Parse } = require('parse/node');
 
-import { getJob, getFunction, resolveError } from '../triggers.js';
-import { maybeRunValidator } from './../cloud-code/Parse.Cloud.Validator.js';
+import { getJob, getFunction, maybeRunValidator, resolveError } from '../triggers.js';
 import PromiseRouter from '../PromiseRouter';
 import { promiseEnforceMasterKeyAccess, promiseEnsureIdempotency } from '../middlewares';
 import { jobStatusHandler } from '../StatusHandler';
@@ -11,9 +10,6 @@ import _ from 'lodash';
 import { logger } from '../logger';
 
 function parseObject(obj) {
-  if (obj === null || obj === undefined) {
-    return obj;
-  }
   if (Array.isArray(obj)) {
     return obj.map(item => parseObject(item));
   }
